Seperated classes Renamed sockclienttest.py to logicclient.py RobotInterface is now named RobotMotion The new robotinterface.py contains socket code Seperated sockets into 2, sender and reciever Improved draw functionality
